fix: delete all previous migrations file, delete existing tables, and execute new migration.
Previous migrations have a problem: at first, we did not use Label.label
as a primary key, so Django automatically create an label_id (integer)
field to be the primary key, and use this in the index of
annotation_label.

After we set Label.label to be the primary key, the label_id in the
index of annotation_label remains type of integer, which conflicts the
current setting: label_id should be label itself, thus is char(50).

In order to fix this, I directly delete existing tables. This can be
done in the beginning phase of a project, but will be painful when we
have already stored a lot of things in the database. Hence it is
necessary to figure out how to fix this kind of problem in a more
appropriate way.

I would like to search the solution for it, however, I have to continue
this project first. So it's probably that I will forget to do this...